requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.15
django-versatileimagefield==3.1
djangorestframework==3.15.2
pillow==10.4.0
python-magic==0.4.27
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
